add: Track_params + requests, little cleanup;

+++ src/app/library/track/service.py
@@ -9,6 +9,7 @@ from fastapi.responses import FileResponse
 
 from ...base.service_base import Service_base
 from .. import models
+from ...rec_service import models as model_params
 from ..track import schemas
 
 
@@ -21,12 +22,14 @@ class Track_service(Service_base):
 
     async def create(self, 
                     schema: schemas.Track_create = Depends(), 
+                    track_params: schemas.Track_params = Depends(),
                     artists: List[int] = Form(...), 
                     album: int = Form(...), 
                     genre: int = Form(...),
                     picture_file: UploadFile = File(...),
                     track_file: UploadFile = File(...),
                     **kwargs) -> Optional[schemas.Track_create]:
+        print(track_params)
         _album = await models.Album.get_or_none(id=album)
         _genre = await models.Genre.get_or_none(id=genre)
         if not _genre:
@@ -36,6 +39,9 @@ class Track_service(Service_base):
         else:
             obj = await self.model.create(**schema.dict(exclude_unset=True), genre=_genre, **kwargs)
         await obj.save()
+        print(track_params.dict(exclude_unset=True))
+        _track_params = await model_params.Track_params.create(**track_params.dict(exclude_unset=True), track_id=obj)
+        await _track_params.save()
         
         picture_file_path = await self.upload_file('track/picture', picture_file)
         if picture_file_path['file_path'] != 'NULL':
@@ -87,6 +93,15 @@ class Track_service(Service_base):
                             media_type=f'audio/{obj.picture_file_path.split(".")[-1]}',
                             filename=f'audio_{obj.id}.{obj.picture_file_path.split(".")[-1]}')
 
+    async def update_track_params(self, 
+                                track_id: int,
+                                params: schemas.Track_params):
+        await model_params.Track_params.filter(track_id=track_id).update(**params.dict(exclude_unset=True))
+
+    async def get_track_params(self, track_id: int):
+        obj = await model_params.Track_params.get(track_id=track_id)
+        return obj
+
     async def change_genre(self, schema, genre, **kwargs) -> Optional[schemas.Track_change_genre]:
         _genre = await models.get(id=genre)
         obj = await self.model.update(schema.dict(exclude_unset=True), genre=_genre, **kwargs)
@@ -110,8 +125,10 @@ class Track_service(Service_base):
             raise HTTPException(
                 status_code=404, detail='Object does not exist'
             )
-        os.remove(obj.track_file_path)
-        os.remove(obj.picture_file_path)
+        if (obj.track_file_path != ''):
+            os.remove(obj.track_file_path)
+        if (obj.picture_file_path != 'data/default_image.png'):
+            os.remove(obj.picture_file_path)
         await obj.delete()
